FIX node build
